messaging_app/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
sqlparse==0.5.3
